+++ src/main/java/insane96mcp/enhancedai/module/mobs/targeting/EAIShulkerNearestAttackTargetGoal.java
@@ -9,8 +9,8 @@ import net.minecraft.world.entity.monster.Shulker;
 import net.minecraft.world.phys.AABB;
 
 public class EAIShulkerNearestAttackTargetGoal<T extends LivingEntity> extends EAINearestAttackableTarget<T> {
-    public EAIShulkerNearestAttackTargetGoal(Shulker goalOwnerIn, Class<T> targetClassIn, boolean checkSight, boolean nearbyOnlyIn, TargetingConditions targetingConditions) {
-        super(goalOwnerIn, targetClassIn, checkSight, nearbyOnlyIn, targetingConditions);
+    public EAIShulkerNearestAttackTargetGoal(Shulker goalOwnerIn, Class<T> targetClassIn, boolean nearbyOnlyIn, TargetingConditions targetingConditions) {
+        super(goalOwnerIn, targetClassIn, nearbyOnlyIn, targetingConditions);
     }
 
     public boolean canUse() {